refactor: repositioning cards
