+++ apps/dashboard/app/components/cards/ResourceGauge.tsx
