+++ apps/dashboard/app/routes/dashboard/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Users } from "lucide-react";
 
-import OverviewCardGrid from "~/components/dashboard/OverviewCardGrid";
-import type { OverviewCardProps } from "~/components/dashboard/OverviewCard";
+import OverviewCardGrid from "~/components/cards/OverviewCardGrid";
+import type { OverviewCardProps } from "~/components/cards/OverviewCard";
 import ServerCardGrid from "~/components/dashboard/ServerCardGrid";
 
 import useDashboard from "./useDashboard";
